[CI/Build] Reorganize V1 tests
